REORG/MINOR: Rename plugin misc.js to utils.js
